+++ HISTORY.md
@@ -3,6 +3,7 @@
 ## SNAPSHOT, v1.1.0
 
 - [hazelcast-camel] Direct implementation of a polling consumer.
+- [hazelcast-camel] Added support for per message destination selection via a message header.
 - [yeti] Improved the performance and error handling of the StompFrameDecoder.
 - [yeti] Implemented maximum frame size in the StompFrameDecoder.